requirements-render-minimal.txt
# ==============================================================================
# RENDER-MINIMAL REQUIREMENTS - Guaranteed to build on free tier
# Ultra-light version that WILL work
# ==============================================================================

# --- Web Server & Framework (REQUIRED) ---
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1

# --- Database & Auth (REQUIRED) ---
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
Flask-Login==0.6.3
Flask-Migrate==4.0.5

# --- Core AI Engine (REQUIRED) ---
google-generativeai>=0.7.0

# --- Configuration & Security (REQUIRED) ---
python-dotenv==1.0.0
Flask-CORS==4.0.0

# --- Document Processing (REQUIRED) ---
PyPDF2==3.0.1
python-docx==1.1.0

# ==============================================================================
# EVERYTHING ELSE COMMENTED OUT FOR NOW
# Uncomment AFTER first successful build
# ==============================================================================

# # --- Additional AI Providers ---
# openai>=1.3.0
# anthropic>=0.7.0
# groq>=0.4.0

# # --- Task Queue ---
# celery==5.3.4
# redis==5.0.1

# # --- Vector Store (HEAVY - causes timeout) ---
# sentence-transformers==2.2.2
# chromadb==0.4.22
# langchain==0.1.0

# # --- Data Processing ---
# pandas==2.1.4
# numpy==1.24.4
# Pillow==10.1.0
